{django_front}[review] Review django_project
Add:
-main pages (show node)
-about pages (unused)
-create node
-sql db with nodes
