+++ DjangoProject/main/admin.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+
+# Register your models here.
+from .models import Node
+
+
+admin.site.register(Node)
