+++ DjangoProject/main/models.py
@@ -1,0 +1,15 @@
+from django.db import models
+
+# Create your models here.
+
+
+class Node(models.Model):
+    title = models.CharField('Название', max_length=50)
+    text = models.TextField('Содержание')
+
+    def __str__(self):
+        return self.title
+
+    class Meta:
+        verbose_name = 'Заметка'
+        verbose_name_plural = 'Заметки'
