+++ DjangoProject/main/templates/main/about.html
@@ -1,0 +1,10 @@
+{% extends 'main/base.html' %}
+
+{% block title %}
+Страница про нас
+{% endblock %}
+
+{% block content %}
+    <h1>Страница о нас</h1>
+    <p1>Страница о нас</p1>
+{% endblock %}
