+++ DjangoProject/main/templates/main/create.html
@@ -1,0 +1,17 @@
+{% extends 'main/base.html' %}
+
+{% block title %}
+Создать заметку
+{% endblock %}
+
+{% block content %}
+    <h1>Создать заметку</h1>
+    <form method="post">
+        {% csrf_token %}
+        {{ form.title }}<br>
+        {{ form.text }}<br>
+        <button type="submit" class="btn btn-success">Добавить</button>
+        <span>{{ error }}</span>
+    </form>
+    
+{% endblock %}
