+++ DjangoProject/main/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+from . import views
+
+
+urlpatterns = [
+    path('', views.index, name='home'),
+    path('about', views.about, name='about'),
+    path('create', views.create, name='create_node'),
+]
